Fix auto-refactored imports
